requirements.txt
mysqlclient
celery==5.2.7
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
ipython==8.10.0
python-dotenv==1.0.0
python-env==1.0.0
redis==4.5.1
requests==2.28.2
Pillow==9.4.0